Completed index page view and separate controller to new file
